+++ app/Http/Controllers/Computer/ComputerController.php
@@ -10,6 +10,7 @@ use App\Models\Computer\Computer;
 use App\Models\Sector\Setor;
 use App\Models\Computer\OperationalSystem;
 use App\User;
+use Illuminate\Support\Facades\Gate;
 
 class ComputerController extends Controller
 {
@@ -20,39 +21,55 @@ class ComputerController extends Controller
 
     public function index()
     {
-        return view('computer.computers');
+        if(Gate::allows('list-computers')){
+            return view('computer.computers');
+        }else{
+            abort(403,'Você não possui permissão para gerenciar computadores. Contate o administrador!');
+        }
     }
 
     public function new()
     {
-        $sectors = Setor::all();
-        $sos = OperationalSystem::all();
-        return view('computer.formcomputer', ['update' => json_encode(false), 'sos' => json_encode($sos),'setores' => json_encode($sectors)]);
+        if(Gate::allows('create-computer')){
+            $sectors = Setor::all();
+            $sos = OperationalSystem::all();
+            return view('computer.formcomputer', ['update' => json_encode(false), 'sos' => json_encode($sos),'setores' => json_encode($sectors)]);
+        }else{
+            abort(403,'Você não possui permissão para adicionar computadores. Contate o administrador!');
+        }
     }
 
     public function view(Request $request){
-        $computer = Computer::find($request->id);
-        $computer->setor = Setor::find($computer->setor_id)->nome;
-        $computer->so = OperationalSystem::find($computer->sistema_operacional_id);
-
-        $articles_computers = ComputerArticle::where('computador_id', $request->id)->get();
-        $assocs = [];
-        foreach ($articles_computers as $item) {
-            $article = Article::where('id_artigo', $item->artigo_id)->get();
-            $article[0]->autor = User::find($article[0]->usuario_autor_id)->nome;
-            $article[0]->contador = $item->contador;
-            $assocs[] = $article[0];
+        if(Gate::allows('view-computer')){
+            $computer = Computer::find($request->id);
+            $computer->setor = Setor::find($computer->setor_id)->nome;
+            $computer->so = OperationalSystem::find($computer->sistema_operacional_id);
+    
+            $articles_computers = ComputerArticle::where('computador_id', $request->id)->get();
+            $assocs = [];
+            foreach ($articles_computers as $item) {
+                $article = Article::where('id_artigo', $item->artigo_id)->get();
+                $article[0]->autor = User::find($article[0]->usuario_autor_id)->nome;
+                $article[0]->contador = $item->contador;
+                $assocs[] = $article[0];
+            }
+    
+            return view('computer.computer', ['computer'=> json_encode($computer), 'articles' => json_encode($assocs)]);
+        }else{
+            abort(403,'Você não possui permissão para vizualizar computadores. Contate o administrador!');
         }
-
-        return view('computer.computer', ['computer'=> json_encode($computer), 'articles' => json_encode($assocs)]);
     }
 
     public function change($id)
     {
-        $sectors = Setor::all();
-        $sos = OperationalSystem::all();
-        $pc = Computer::find($id);
-        return view('computer.formcomputer', ['update' => json_encode(true), 'sos' => json_encode($sos),'setores' => json_encode($sectors), 'computerUpdate' => json_encode($pc)]);
+        if(Gate::allows('view-computer')){
+            $sectors = Setor::all();
+            $sos = OperationalSystem::all();
+            $pc = Computer::find($id);
+            return view('computer.formcomputer', ['update' => json_encode(true), 'sos' => json_encode($sos),'setores' => json_encode($sectors), 'computerUpdate' => json_encode($pc)]);
+        }else{
+            abort(403,'Você não possui permissão para editar computadores. Contate o administrador!');
+        }
     }
 
     public function getComputers()
@@ -69,6 +86,10 @@ class ComputerController extends Controller
 
     public function create(Request $request)
     {
+        if (Gate::denies('create-computer')) {
+            return false;
+        }
+
         $computer = new Computer();
         $computer->placa_mae = $request->placa_mae;
         $computer->processador = $request->processador;
@@ -98,6 +119,10 @@ class ComputerController extends Controller
 
     public function update(Request $request)
     {
+        if (Gate::denies('edit-computer')) {
+            return false;
+        }
+
         $computer = Computer::find($request->id_computador);
         $computer->placa_mae = $request->placa_mae;
         $computer->processador = $request->processador;
@@ -126,6 +151,10 @@ class ComputerController extends Controller
 
     public function delete($id)
     {
+        if (Gate::denies('delete-computer')) {
+            return false;
+        }
+
         $computer = Computer::find($id);
         
         if($computer->delete()){
